pyproject.toml
[project]
name = "sistem-store"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.13.1",
    "blinker==1.7.0",
    "chardet==5.2.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "flask-login==0.6.3",
    "flask-migrate==4.0.7",
    "flask-sqlalchemy==3.1.1",
    "flask==3.0.0",
    "greenlet==3.0.3",
    "gunicorn==22.0.0",
    "itsdangerous==2.1.2",
    "jinja2==3.1.2",
    "mako==1.3.5",
    "markupsafe==2.1.3",
    "mysql==0.0.3",
    "mysqlclient==2.2.4",
    "packaging==24.0",
    "pillow==10.4.0",
    "python-dotenv==1.0.1",
    "reportlab==4.2.2",
    "sqlalchemy==2.0.30",
    "typing-extensions==4.11.0",
    "werkzeug==3.0.1",
]
